requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Web scraping
playwright==1.40.0

# AI/ML
openai==1.3.7
tiktoken==0.5.2

# Google Sheets integration
gspread==5.12.0
oauth2client==4.1.3

# PDF processing
PyMuPDF==1.23.8

# Utilities
requests==2.31.0
python-multipart==0.0.6 
redis==5.0.4

# Resilience and Anti-Bot Features
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3
asyncio-throttle==1.0.2
tenacity==8.2.3 